pyproject.toml
[tool.poetry]
name = "assignment2"
version = "0.1.0"
description = ""
authors = ["aish"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
